+++ R/as_draws.R
@@ -16,9 +16,13 @@
 #'  * `beta`\cr Time-invariant regression coefficients.
 #'  * `delta`\cr Time-varying regression coefficients.
 #'  * `nu`\cr Random intercepts.
+#'  * `psi`\cr Latent factor.
+#'  * `lambda`\cr Factor loadings.
 #'  * `tau`\cr Standard deviations of the spline coefficients of `delta`.
 #'  * `tau_alpha`\cr Standard deviations of the spline coefficients of
 #'    time-varying `alpha`.
+#'  * `tau_psi`\cr Standard deviations of the spline coefficients of
+#'     latent factor `psi`.
 #'  * `sigma_nu`\cr Standard deviation of the random intercepts `nu`.
 #'  * `sigma`\cr Standard deviations of gaussian responses.
 #'  * `phi`\cr Dispersion parameters of negative binomial responses.