+++ R/stanblocks.R
@@ -44,7 +44,8 @@ create_functions <- function(dformula, idt, vars) {
 #' @noRd
 create_data <- function(dformula, idt, vars) {
   has_splines <- any(unlist(lapply(vars, "[[", "has_varying"))) ||
-    any(unlist(lapply(vars, "[[", "has_varying_intercept")))
+    any(unlist(lapply(vars, "[[", "has_varying_intercept"))) ||
+    any(unlist(lapply(vars, "[[", "has_lfactor")))
   mtext <- paste_rows(
     "int<lower=1> T; // number of time points",
     "int<lower=1> N; // number of individuals",
@@ -54,6 +55,7 @@ create_data <- function(dformula, idt, vars) {
     onlyif(has_splines, "int<lower=1> D; // number of B-splines"),
     onlyif(has_splines, "matrix[D, T] Bs; // B-spline basis matrix"),
     "int<lower=0> M; // number of channels with random intercept",
+    "int<lower=0> P; // number of channels with latent factor",
     .indent = idt(1),
     .parse = FALSE
   )
@@ -82,12 +84,28 @@ create_transformed_data <- function(dformula, idt, vars) {
     declarations[i] <- tr_data$declarations
     statements[i] <- tr_data$statements
   }
+
+  declare_AQR <- "matrix[N, N - 1] A_qr;"
+  state_AQR <- paste_rows(
+    "{",
+    "matrix[N, N] A = diag_matrix(rep_vector(1, N));",
+    "for (i in 1:K - 1) A[K, i] = -1;",
+    "A[K, K] = 0;",
+    "A_qr = qr_Q(A)[ , 1:(N - 1)];",
+    "}",
+    .indent = idt(c(0, 2, 2, 2, 2, 1)),
+    .parse = FALSE
+  )
+  has_lfactor <- length(attr(dformula, "lfactor")$responses) > 0L
   paste_rows(
     "transformed data {",
-    "  // Parameters for vectorized priors",
+    "// Parameters for vectorized priors",
     declarations,
+    onlyif(has_lfactor, declare_AQR),
     statements,
+    onlyif(has_lfactor, state_AQR),
     "}",
+    .indent = idt(c(0, 1, 0, 1, 0, 1, 0)),
     .parse = FALSE
   )
 }
@@ -124,19 +142,39 @@ create_parameters <- function(dformula, idt, vars) {
       .indent = idt(c(1, 1, 1))
     )
   )
+
+  lfactortext <- ifelse_(
+    identical(length(attr(dformula, "lfactor")$responses), 0L),
+    "",
+    paste_rows(
+      "// Latent factor splines",
+      onlyif(
+        attr(dformula, "lfactor")$correlated,
+        "cholesky_factor_corr[P] L_lf; // Cholesky for correlated factors"
+      ),
+      ifelse_(
+        attr(dformula, "lfactor")$noncentered_psi,
+        "matrix[P, D - 1] omega_raw_psi;", "matrix[P, D] omega_raw_psi;"
+      ),
+      .indent = idt(c(1, 1, 1))
+    )
+  )
+
   pars <- character(length(dformula))
   for (i in seq_along(dformula)) {
     family <- dformula[[i]]$family$name
     line_args <- c(list(y = vars[[i]]$resp, idt = idt), vars[[i]])
     pars[i] <- lines_wrap("parameters", family, line_args)
   }
-  paste_rows("parameters {", splinetext, randomtext, pars, "}", .parse = FALSE)
+  paste_rows("parameters {", splinetext, randomtext, lfactortext, pars, "}",
+    .parse = FALSE)
 }
 
 #' @describeIn create_function Create the 'Transformed Parameters'
 #'   Block of the Stan Model Code
 #' @noRd
 create_transformed_parameters <- function(dformula, idt, vars) {
+
   randomtext <- ""
   nus <- attr(dformula, "random")$responses
   M <- length(nus)
@@ -162,6 +200,41 @@ create_transformed_parameters <- function(dformula, idt, vars) {
         )
     }
   }
+
+
+  lfactortext <- ""
+  psis <- attr(dformula, "lfactor")$responses
+  P <- length(psis)
+  if (P > 0) {
+    if (attr(dformula, "lfactor")$noncentered_psi) {
+      tau_psi <- ifelse(attr(dformula, "lfactor")$nonzero_lambda,
+        paste0("tau_psi_", psis, " * "), "")
+      # create noise terms for latent factors
+      lfactortext <- ifelse_(
+        attr(dformula, "lfactor")$correlated,
+        paste_rows(
+          "matrix[P, D - 1] omega_psi = L_lf * omega_raw_psi;",
+          glue::glue(
+            "row_vector[D] omega_psi_{psis} = \\
+            append_col(0, {tau_psi}omega_psi[{1:P}, ]);"),
+          .indent = idt(1)
+        ),
+        paste_rows(
+          glue::glue("row_vector[D] omega_psi_{psis} = \\
+            append_col(0, {tau_psi}omega_raw_psi[{1:P}, ]);"),
+          .indent = idt(1)
+        )
+      )
+    } else {
+      lfactortext <-
+        paste_rows(
+          glue::glue("row_vector[D] omega_psi_{psis} = \\
+            omega_raw_psi[{1:P}, ];"),
+          .indent = idt(1)
+        )
+    }
+  }
+
   n <- length(dformula)
   declarations <- character(n)
   statements <- character(n)
@@ -175,6 +248,7 @@ create_transformed_parameters <- function(dformula, idt, vars) {
   paste_rows(
     "transformed parameters {",
     randomtext,
+    lfactortext,
     declarations,
     statements,
     "}",
@@ -207,7 +281,7 @@ create_model <- function(dformula, idt, vars, backend) {
       } else {
         randomtext <- paste_rows(
           "nu_raw ~ multi_normal_cholesky(0, diag_pre_multiply(sigma_nu, L));",
-          onlyif(attr(dformula, "random")$correlated, "L ~ {L_prior};"),
+          "L ~ {L_prior};",
           .indent = idt(c(1, 1))
         )
       }
@@ -227,6 +301,77 @@ create_model <- function(dformula, idt, vars, backend) {
       }
     }
   }
+
+  lfactortext <- ""
+  psis <- attr(dformula, "lfactor")$responses
+  P <- length(psis)
+  if (P > 0) {
+    if (attr(dformula, "lfactor")$correlated) {
+      L_prior <- attr(vars, "common_priors")
+      L_prior <- L_prior[L_prior$parameter == "L_lf", "prior"]
+      if (attr(dformula, "lfactor")$noncentered_psi) {
+        lfactortext <- paste_rows(
+          "to_vector(omega_raw_psi) ~ std_normal();",
+          "L_lf ~ {L_prior};",
+          .indent = idt(c(1, 1))
+        )
+      } else {
+        if (any(attr(dformula, "lfactor")$nonzero_lambda)) {
+          tau <- paste0(ifelse(attr(dformula, "lfactor")$nonzero_lambda,
+            paste0("tau_psi_", psis), "1"), collapse = ", ")
+          lfactortext <- paste_rows(
+            "L_lf ~ {L_prior};",
+            "{{",
+            "vector[P] tau_psi = [{tau}]';",
+            "matrix[P, P] Ltau = diag_pre_multiply(tau_psi, L_lf);",
+            "for (i in 2:D) {{",
+            "omega_raw_psi[, i] ~ multi_normal_cholesky(omega_raw_psi[, i - 1], Ltau);",
+            "}}",
+            "}}",
+            .indent = idt(c(1, 1, 2, 2, 2, 3, 2, 1))
+          )
+        } else {
+          lfactortext <- paste_rows(
+            "L_lf ~ {L_prior};",
+            "for (i in 2:D) {{",
+            "omega_raw_psi[, i] ~ multi_normal_cholesky(omega_raw_psi[, i - 1], L_lf);",
+            "}}",
+            .indent = idt(c(1, 1, 2, 1))
+          )
+        }
+      }
+    } else {
+      if (attr(dformula, "lfactor")$noncentered_psi) {
+        lfactortext <- paste_rows(
+          "to_vector(omega_raw_psi) ~ std_normal();",
+          .indent = idt(1)
+        )
+      } else {
+        if (any(attr(dformula, "lfactor")$nonzero_lambda)) {
+          psis <- attr(dformula, "lfactor")$responses
+          P <- length(psis)
+          tau <- paste0(ifelse(attr(dformula, "lfactor")$nonzero_lambda,
+            paste0("tau_psi_", psis), "1"), collapse = ", ")
+          lfactortext <- paste_rows(
+            "{{",
+            "vector[P] tau_psi = [{tau}]';",
+            "for (i in 2:D) {{",
+            "omega_raw_psi[, i] ~ normal(omega_raw_psi[, i - 1], tau_psi);",
+            "}}",
+            "}}",
+            .indent = idt(c(1, 2, 2, 3, 2, 1))
+          )
+        } else {
+          lfactortext <- paste_rows(
+            "for (i in 2:D) {{",
+            "omega_raw_psi[, i] ~ normal(omega_raw_psi[, i - 1], 1);",
+            "}}",
+            .indent = idt(c(1, 2, 1))
+          )
+        }
+      }
+    }
+  }
   mod <- character(length(dformula))
   for (i in seq_along(dformula)) {
     family <- dformula[[i]]$family$name
@@ -236,18 +381,18 @@ create_model <- function(dformula, idt, vars, backend) {
     )
     mod[i] <- lines_wrap("model", family, line_args)
   }
-  paste_rows("model {", splinetext, randomtext, mod, "}", .parse = FALSE)
+  paste_rows("model {", splinetext, randomtext, lfactortext, mod, "}", .parse = FALSE)
 }
 
 #' @describeIn create_function Create the 'Generated Quantities'
 #'   Block of the Stan Model Code
 #' @noRd
 create_generated_quantities <- function(dformula, idt, vars) {
-  gen <- ""
+  gen_nu <- ""
   M <- length(attr(dformula, "random")$responses)
   if (M > 0 && attr(dformula, "random")$correlated) {
     # evaluate number of corrs to avoid Stan warning about integer division
-    gen <- paste_rows(
+    gen_nu <- paste_rows(
       "corr_matrix[M] corr_matrix_nu = multiply_lower_tri_self_transpose(L);",
       "vector<lower=-1,upper=1>[{(M * (M - 1L)) %/% 2L}] corr_nu;",
       "for (k in 1:M) {{",
@@ -258,8 +403,27 @@ create_generated_quantities <- function(dformula, idt, vars) {
       .indent = idt(c(1, 1, 1, 2, 3, 2, 1))
     )
   }
-  if (any(nzchar(gen))) {
-    paste_rows("generated quantities {", gen, "}", .parse = FALSE)
+  gen_psi <- ""
+  P <- length(attr(dformula, "lfactor")$responses)
+  if (P > 0 && attr(dformula, "lfactor")$correlated) {
+    # evaluate number of corrs to avoid Stan warning about integer division
+    tau <- paste0(ifelse(attr(dformula, "lfactor")$nonzero_lambda,
+      paste0("tau_psi_", psis), "1"), collapse = ", ")
+    gen_psi <- paste_rows(
+      "vector[P] tau_psi = [{tau}]';",
+      "corr_matrix[P] corr_matrix_psi =
+       multiply_lower_tri_self_transpose(diag_pre_multiply(tau_psi, L_lf));",
+      "vector<lower=-1,upper=1>[{(P * (P - 1L)) %/% 2L}] corr_psi;",
+      "for (k in 1:P) {{",
+      "for (j in 1:(k - 1)) {{",
+      "corr_psi[choose(k - 1, 2) + j] = corr_matrix_psi[j, k];",
+      "}}",
+      "}}",
+      .indent = idt(c(1, 1, 1, 1, 2, 3, 2, 1))
+    )
+  }
+  if (any(nzchar(gen_nu)) || any(nzchar(gen_psi))) {
+    paste_rows("generated quantities {", gen_nu, gen_psi, "}", .parse = FALSE)
   } else {
     NULL
   }