+++ NEWS.md
@@ -27,6 +27,7 @@
   * Breaking change: The shrinkage parameter which was previously named as 
     `lambda` is now `xi` in order to free `lambda` for factor loadings parameter 
      as is customary in factor analysis.
+  * Added support for correlated latent dynamic factors (modelled as splines).
 
 # dynamite 0.0.2